+++ public/assets/images/icons/projectV2/PointMarkerIconsSymbol.tsx
